Add files via upload
These files are in regards to the individual assignment for Trinity College Dublin Financial Analytics (BU7152) for the Spring term 2022. The Q2/Q3 notebooks answer the questions from the projects document.

These particular files look into factor analysis of cryptocurrencies, and the Q2 looks into herding for a select group of stocks in the current market conditions over a defined timeframe.